feat: deep refactoring BREAKING CHANGE: OIDC configuration changed. Only discovery URL supported now
