+++ mlflow_oidc_auth/__init__.py
@@ -1,5 +1,5 @@
 import os
 
-version = os.environ.get("MLFLOW_OIDC_AUTH_VERSION", "2.0.0.dev0")
+version = os.environ.get("MLFLOW_OIDC_AUTH_VERSION", "3.0.0.dev0")
 
 __version__ = version
