+++ mlflow_oidc_auth/store.py
@@ -1,5 +1,5 @@
 from mlflow_oidc_auth.sqlalchemy_store import SqlAlchemyStore
-from mlflow_oidc_auth.config import AppConfig
+from mlflow_oidc_auth.config import config
 
 store = SqlAlchemyStore()
-store.init_db((AppConfig.get_property("OIDC_USERS_DB_URI")))
+store.init_db(config.OIDC_USERS_DB_URI)
