+++ mlflow_oidc_auth/utils.py
@@ -2,6 +2,11 @@ from flask import request, session
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import BAD_REQUEST, INVALID_PARAMETER_VALUE
 from mlflow.server import app
+from mlflow.server.handlers import _get_tracking_store
+
+from mlflow_oidc_auth.store import store
+from mlflow_oidc_auth.auth import validate_token
+
 
 def get_request_param(param: str) -> str:
     if request.method == "GET":
@@ -25,6 +30,7 @@ def get_request_param(param: str) -> str:
         )
     return args[param]
 
+
 def get_username():
     username = session.get("username")
     if username:
@@ -35,5 +41,31 @@ def get_username():
             app.logger.debug(f"Username from basic auth: {request.authorization.username}")
             return request.authorization.username
         if request.authorization.type == "bearer":
-            return request.authorization.email
+            username = validate_token(request.authorization.token).get("email")
+            app.logger.debug(f"Username from bearer token: {username}")
+            return username
     return None
+
+
+def get_is_admin() -> bool:
+    return bool(store.get_user(get_username()).is_admin)
+
+
+def get_experiment_id() -> str:
+    if request.method == "GET":
+        args = request.args
+    elif request.method in ("POST", "PATCH", "DELETE"):
+        args = request.json
+    else:
+        raise MlflowException(
+            f"Unsupported HTTP method '{request.method}'",
+            BAD_REQUEST,
+        )
+    if "experiment_id" in args:
+        return args["experiment_id"]
+    elif "experiment_name" in args:
+        return _get_tracking_store().get_experiment_by_name(args["experiment_name"]).experiment_id
+    raise MlflowException(
+        "Either 'experiment_id' or 'experiment_name' must be provided in the request data.",
+        INVALID_PARAMETER_VALUE,
+    )
